cambios a mejor en el ejercicios chatroom
